Display number of devices and registration limit on the sidebar
As a quick feedback for the user, display always visible fields on the
sidebar to report:
- The number of connected devices, over the total number of devices.
- The number of registered devices, over the total number of devices
that can be registered if a device registration limit exists.
